sideload script: simplify theme.json requirements
All theme.json files had "enabled": false as a workaround. This was an
ok fix when we needed to deal with using grep and sed and the like.
This meant that we overwrote the "enabled" value as well each time.

With some experimentation, most of the fields in theme.json are not
needed for sideloading. Removed them.